friendrequests/serializers.py
from rest_framework import serializers
from .models import FriendRequest
from profiles.models import User
class FriendRequestSerializer(serializers.ModelSerializer):
    class Meta:
        model = FriendRequest
        fields = ['id', 'sender', 'receiver', 'status', 'created_at']

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'name', 'email', 'bio', 'is_active', 'created_at', 'updated_at')
